refactor: infra view provider class
